+++ rfs/nostr-blossom/src/client.rs
@@ -4,11 +4,9 @@ use std::time::Duration;
 
 use base64::Engine;
 use base64::engine::general_purpose;
-use nostr::event::FinalizeEventAsync;
 use nostr::hashes::Hash;
 use nostr::hashes::sha256::Hash as Sha256Hash;
-use nostr::signer::{AsyncGetPublicKey, AsyncSignEvent};
-use nostr::{Event, EventBuilder, PublicKey, Timestamp, Url};
+use nostr::prelude::*;
 use reqwest::header::{AUTHORIZATION, CONTENT_TYPE, HeaderMap, HeaderValue, RANGE};
 #[cfg(not(target_arch = "wasm32"))]
 use reqwest::redirect::Policy;